Update scripts to handle updated entities
