+++ scripts/scripts/__init__.py
@@ -1,23 +1,19 @@
 from . import questions
-from . import insert_into_db
 from . import sentences
 from . import select_questions
-from . import db_dump
 from . import tts
 from . import upload_to_r2
 from . import explanations
-from . import update_question_data
 from . import insert_sentences_into_db
+from . import insert_questions_into_db
 
 __all__ = [
     "sentences",
     "explanations",
     "questions",
     "select_questions",
-    "insert_into_db",
     "tts",
     "upload_to_r2",
-    "db_dump",
-    "update_question_data",
     "insert_sentences_into_db",
+    "insert_questions_into_db",
 ]
